+++ apps/prairielearn/src/pages/instructorInstanceAdminPublishing/components/CourseInstancePublishing.tsx
@@ -1,18 +1,17 @@
 import { QueryClient } from '@tanstack/react-query';
+import { useState } from 'preact/compat';
 import { FormProvider, useForm } from 'react-hook-form';
 
+import {
+  CourseInstancePublishingForm,
+  type PublishingFormValues,
+} from '../../../components/CourseInstancePublishingForm.js';
 import type { StaffCourseInstance } from '../../../lib/client/safe-db-types.js';
 import { QueryClientProviderDebug } from '../../../lib/client/tanstackQuery.js';
 import type { CourseInstancePublishingExtensionWithUsers } from '../instructorInstanceAdminPublishing.types.js';
 import { dateToPlainDateTime } from '../utils/dateUtils.js';
 
-import {
-  CourseInstancePublishingForm,
-  type PublishingFormValues,
-} from './CourseInstancePublishingForm.js';
 import { PublishingExtensions } from './PublishingExtensions.js';
-
-const queryClient = new QueryClient();
 
 export function CourseInstancePublishing({
   courseInstance,
@@ -29,6 +28,8 @@ export function CourseInstancePublishing({
   publishingExtensions: CourseInstancePublishingExtensionWithUsers[];
   isDevMode: boolean;
 }) {
+  const [queryClient] = useState(() => new QueryClient());
+
   const originalStartDate = courseInstance.publishing_start_date;
   const originalEndDate = courseInstance.publishing_end_date;
 
@@ -81,7 +82,7 @@ export function CourseInstancePublishing({
 
           <FormProvider {...methods}>
             <CourseInstancePublishingForm
-              courseInstance={courseInstance}
+              displayTimezone={courseInstance.display_timezone}
               canEdit={canEdit}
               originalStartDate={courseInstance.publishing_start_date}
               originalEndDate={courseInstance.publishing_end_date}